Handle Xen code configured into modern kernels
When building a current kernel with the in-kernel xen code enabled on
amd64, kernel-package does not recognize this correctly because it
matches on a config variable which apparently is no longer set (or
maybe even was never used in the in-kernel xen code). This commit adds
an extra stanza that adds support for the Xen in new kernels. The
suggestion and bug report were from  Michael Holzt.

Closes: #503686

Signed-off-by: Manoj Srivastava <srivasta@debian.org>