Pass along the version on the command line when using the minimal rules
When the ./debian directory does not exist, the frirs step is to use a
minimal rules file to create the ./debian directory, and _then_ use the
full rules file for all other actions. Unfortunately, the minimal rules
file did not know what version of kernel-package it belonged to, so now
make-kpkg provides that indormation.

Closes: Bug#508198

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>