Get rid of the facility to patch kernel sources
The patch the kernel facility was adding complexity, and failing to
provide the flexibility required for a generic patching facility. It used
to be useful at one point, but in the modern parlance, witht he
widespread use of distribute version control systems, and various
facilities to manage source and patch them, the built in version was
clunky.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>