Remove special case code for official kernels
For the longest tine (well,  ever since Herbert Xu too over building
kernel images from me), kernel-package has carried specal case code for
official images. This has caused some problems, recently, since the need
to preserve ./debian has caused no end of problems when the version
changed out from under ./debian, or when people wanted to edit a file and
expected kernel-package to do a minimal recompile.

However, sometime in the Etch release cycle, the kernel team deprecated
kernel-package as the means of building official kernels, and therefore,
a full release cycle later, we can get rid of the special case rules used
for official packages. Also, this allows us to drop ./debian at teh drop
of a hart, and recreate it with an version that reflects the current
state of the kernel sources. If people want to dump files into ./debian,
we can come up with a different scheme to allow that, if such need can be
demonstrated.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>