Modified the maintainer scripts for the header package
This patch extends the header package maintainer scripts to all four
actions, and creates the directories that the hook scripts can be
dropped in for site configuration of the actions that the package
takes. There are tests that do basic syntax checks on the new scripts
introduced.  Most of the changes in this patch are related to the new
maintainer scripts introduced.

The link manipulation that the header packages used to do in the
postinst has been removed, and is now provided as an example
script.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>