Modified the maintainer scripts for the doc package
This patch extends the doc package maintainer scripts to all four
actions, and creates the directories that the hook scripts can be
dropped in for site configuration of the actions that the package
takes. There are tests that do basic syntax checks on the new scripts
introduced.  Most of the changes in this patch are related to the new
maintainer scripts introduced. There should be no user visible changes
introduced beyong that.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>