Image postinst no longer runs a boot loader
Now that we have a mechanism for running arbitrary scripts when the
image packages are manipulated, we can stop embedding the boot loader
actions in the package itself.  This means that lilo, elilo, etc will
no longer be run directly by the post isnt, and all the code related
to detecting the boot loader, managing the configuration, and adding
bits about bootloader documentation is all removed from the postinst.

This allows the image package to be more flexible, since the end user
is no longer restricted to the actions encoded in the image
package. This is a fairly large change.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>