The postinst no longer manipulates symlinks
This is a shift from previous behaviour. Any symbolic link manipulation
must now be done with hook scripts in /etc/kernel/*.d directories.

Firstly, modern boot loaders scan the boot directory for kernel images,
and the user no longer has to code in the path to the symbolic links that
the kernel image package used to manipulate.

Secondly, hardcoding the behaviour into the postinst made for a very
rigid policy; and user wanted more flexibility than that. There is an
example shipped with the package that shows a more flexible scheme that
kept two symbolic links for version 2.4 kernels, and two symbolic links
for 2.6 kernels; it can be easily modified to keep two links for 2.9
krenels and two links for 2.8 kernels, or one of each, or whatever the
user wants.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>