make-kpkg removes and re-creates ./debian on every invocation
This does make the kernel-package far more nimble; we now offer less
surprise to users who did not expect stampts that the kernel-packagge
used to not do duplicate work. Now, if you edit a couple of files in the
kernel source, and run make-kpkg, the kernel will build as
expected. There are no more "version mismatch" errors, and the kernel
version can be modified using localconfig as one desires. With this,
kernel-package can rountinely be used to build kernels out of the git
tree.

The con is that we no longer cater to official kernels, or to anyone who
expected content in ./debian to persist. At some point, there are plans
to implement an overlay directory that will shadow
/usr/share/kernel-package/ruleset, but that is not yet implemented.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>