Add a "debuginfo" package
This commit adds support for creating a kernel package which contains
the Linux kernel debug image. This means the debugging information for
the modules in the kernel  image package, and the uncompressed vmlinux
image. This builds on suggestions and code from Troy Heber, Theodore
Y. Ts'o, and Dann Frazier.

As support for kexec/kdump support becomes more real, it will be very
useful to be able to build kernels that have debugging information
available, but not necessarily to install the debugging information on
every single client system.

The .ko object files are stripped of the debugging information before
they are placed in the standard linux-image installation package.
However, before that, a copy of the vmlinux and unstripped .ko files
are saved in a "debuginfo" package.  This means that if someone ends
up with a crash dump, they can send it to a support engineer and only
the support engineer needs to install the debuginfo package and use
the "crash" utility to debug the crash dump.  It's also useful for
developers, since the debuginfo information can be stored somewhere
outside of /lib for storing its debug information, for example.  This
is useful for keeping the size of the root partition small, for those
who keep / separate from /usr.

The locations used are compatible also with SystemTap, which provides
free software infrastructure to simplify the gathering of information
about the running Linux system. This assists diagnosis of a
performance or functional problem. SystemTap eliminates the need for
the developer to go through the tedious and disruptive instrument,
recompile, install, and reboot sequence that may be otherwise required
to collect data.

Closes: #365349, #517290, #373639

Signed-off-by: Manoj Srivastava <srivasta@debian.org>