Export maintainer script parameters to hook scripts
Maintainer script option need to be passed on to the hook scripts,
since there are many ways that the script may berun (for example,
postinst hooks scriopts are called even when the installation is
aborted). This commit adds the maintainer script arguments to an
envronment variable, DEB_MAINT_PARAMS.

Updated the example hook scripts shipped in the documentation to
demonstrate how these arguments can be accessed and used.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>