Keep modifications to the upstream directory to a minimum
When upstream kernel build system started delting ./debian directory on
clean, we had to start taking making an effort to preserve ./debian,
since there was a lot of information in there that was important to
us. So we had a way of diverting the makefile and scripts away that did
the removal of ./debian, and restoring them when we wanted to.

Now that the ./debian directory has become ephemeral, we only need to
preserve it when calling clean during the buildpackage target, since
invoking make-kpkg at any other time cleans ou ./debian anyway. So we can
be more selecting in when we divert upstream scripts; and thus leave the
upstream directory more like we found it (this makes git happy).

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>