Added a new feature: overlay directory
This commits add an overlay feature. The specified directory should
contain files that will be placed in the ./debian directory of the kernel
sources, in preparation to building the debian packages. The files will
replace anything in /usr/share/kernel-package that would normally be
placed there, and it is up to the user to make sure that the files in the
overlay directory are compatible with make-kpkg.  If you break make-kpkg
with an overlay file, you get to keep the pieces. The same effect can be
achieved by setting the environment variable KPKG_OVERLAY_DIR.

Please  note  that  overlay-dir/Control  and  overlay-dir/changelog are
special, and variable substitution is performed on these files. Use
/usr/share/kernel-package/Control and /usr/share/kernel-package/changelog
files as templates.

If a overlay-dir/post-install executable (or executable script) exists,
it shall be run immediately after ./debian is populated.  The  script
shall be executed in the ./debian directory. This can be used, For
instance, to delete files the user does not want, or to take actions
other than simple replacement.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>