Move invocation of postinst hook scripts after /etc/kernel/*.d/* scripts
A lot of people have update-grub in their /etc/kernel-img.conf
postint_hook setting. If that is called early, the initramfs invocation
in /etc/kernel/postinst.d/initramfs has not yet happened, so the update
grub script does not notice the initramfs, making the system unbootable.

Yes, the update grub cant be moved into /etc/kernel/postinst.d/ too, but
this is friendlier.

Closes: #523931

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>