Clarify that kernel-package no longer will create initrds
Instead, now all that --initrd does is to arrange to convey to the hook
scripts that this image requires an initrd, and that the initrd
generation hook scripts should not short circuit early. Without this
option, the example initramfs hook scripts bundled in with kernel-package
will take no action on installation.

Closes: #524499

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>