Take a first step to being able to build alternate targets
Added variable that define targets are used in in the default build in
ruleset/targets/{common,image}.mk to build and install the default
target (usually the kernel image and modules). Change this at your
peril. since the targets called isntead of these might subtly break the
rest of the kerel package building. But, at least, this is a start to
being able to build alternate targets, as long as theya re drop in
replacements as far as packaging is concerned.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>