Remove mention of removed feature of on-the-fly patching
The feature of patching the kernel while compiling it has been removed,
since it was way more trouble than it was worth, did not play nice with a
git repo, and has a  n easy work around -- just pre-patching the
source. This commit removes mentioning it as a supported feature.

Closes: #552521

Signed-off-by: Manoj Srivastava <srivasta@debian.org>