Rework handling of XEN packages
So far, kernel-package would pay special attention to the
configuration files that enabled XEN hypervisor/guest images, mangling
the package name, and changing the file paths where the images were
dumped. This caused no end of confusion to third party tools, and also
surprised people when packages with different names had unexpected
file conflicts. With dpkg defaulting to --force-overwrite that could
be catastrophic. This is because a normal kernel image, a XenU or a
Xen0 image, all looked for modules in /lib/modules/$version, and would
conflict.

With this commit, kernel-package stops paying attention to the XEN
configuration hints in the config, producing a normal kernel image. If
users want to have a DomU or a Dom0 image from the same kernel
version, they will have to do what people wanting to run kernels witht
he same version with differing configurations have always had to do:
use --append-to-version to make the kernel images packages not
conflict.

Closes: #556929

Signed-off-by: Manoj Srivastava <srivasta@debian.org>