[master]: allow arch specific Makefiles to override objcopy as needed
kernel-package fails to build a kernel package on hppa when the target kernel
is 64-bit. It fails because kernel-package uses hardcoded "objcopy" to build
/usr/lib/debug debuginfo files for kernel modules but 64-bit kernel needs
crosscompile prefix on parisc and normal objcopy does not understand the 64-bit
module format.

This change uses the same mechanism as the kernel Makefiles to allow
arch specific Makefiles to specify what to use as objcopy. Thanks to
Meelis Roos <mroos@linux.ee>

Closes: Bug#576725

Signed-off-by: Manoj Srivastava <srivasta@debian.org>