[master]: A new bug fixing release.
The fix for building packages with the correct sub--architecture was
flawed, in the sense it always looked at theKPKG_ARCH, even when there
was no need to stray away from DEB_HOST_ARCH. This commit relieas on
dpkg-architecture to tell us what the proper arch should be for a
situation where we are building for a subarchitecture.
(Closes: #748316).

Signed-off-by: Manoj Srivastava <srivasta@golden-gryphon.com>