Root cause failure to build debs
Firstly, this is not true in general, and I could not reproduce it as
long as I use fakeroot.

Bug fix: "cannot produce debs anymore", thanks to Ritesh Raj Sarraf
The real issue here was that the kernel was built using the deprecated
sudo, instead of fakeroot, and created ./debian directory as
root. Clean, however, is not run as root, so it could not delete
./debian, and failed. This is now documented. (Closes: #812470).

Signed-off-by: Manoj Srivastava <srivasta@debian.org>